requirements.txt
scipy==1.2.1
dustpy.egg==info
corner==2.0.1
h5py==2.9.0
numpy==1.16.2
seaborn==0.9.0
astropy==3.0.5
matplotlib==3.0.2
emcee==2.2.1
pandas==0.23.4
dsharp_opac==1.1.4

